added alternative method using new MediaRecorder in browser
This way the browser does the encoding, not ffmpeg compiled to javascript. At
the moment this only works with Firefox Developer Edition 47.

It can crop and scale the video through a canvas and then captureStream on the
canvas element. More video editing can happen there. The video section is
selected by pressing start and stop recording. Audio is recorded by getting a
stream from directly from the video element with mozCaptureStream. Standard
captureStream isn't supported yet. The audio stream is routed through the Audio
API without applying any change. The stream is duplicated to get playback and
recording. When recording is stopped, a download is generated at the bottom of
the page.

The preview canvas below the video element is updated all the time with
requestAnimationFrame.