made audio capture optional if no video.captureStream
Only firefox supports video.mozCaptureStream (with prefix), e.g., chromium does
not. The video capture goes through a canvas element and works in both browsers
but for audio capture, I have not found any other way than using
video.captureStream as source for the Audio API.

So, audio is disabled if the feature is not available.
TODO: Add indicator if audio is available and also make it possible to disable
audio by user request.